Added ShortcutUtilities project and Program.cs
Added a new project `ShortcutUtilities` to the solution file `UtilitiesForAlibre.sln`. The project configurations for Debug and Release have been set up. A new file `Program.cs` was added to the `ShortcutUtilities` project. This file contains a `Program` class with a `Main` method that provides a console interface for the user to run methods from the `KeyboardShortcutGenerator` class. The `KeyboardShortcutGenerator` class generates all possible combinations of keyboard shortcuts and can print them to the console or save them to a text file. A new project file `ShortcutUtilities.csproj` has been added, targeting .NET Framework 4.8.1 and .NET 8.0, using C# 9.0, and built for the x64 platform. The output type of the project is an executable (Exe).